build: upgrade delta-kernel-rs from v0.17.0 to v0.21.0

+++ src/DeltaLake/Kernel/Callbacks/Allocators/StringAllocatorCallbacks.cs
@@ -39,7 +39,7 @@ namespace DeltaLake.Kernel.Callbacks.Allocators
             //
             for (int i = 0; i < len; i++)
             {
-                unmanagedMemory[i] = slice.ptr[i];
+                unmanagedMemory[i] = (byte)slice.ptr[i];
             }
 
             // Set the null terminator.
